requirements.txt
gradio
hydra-core
openai